+++ src/components/settings/GeneralSettings.tsx
@@ -129,6 +129,16 @@ export default function GeneralSettings(): ReactElement {
 
         <div className='flex justify-between items-center'>
           <div className='flex flex-col gap-2 w-1/2'>
+            <p className='text-sm text-content font-bold'>{t('settings.general.closeToTray')}</p>
+            <p className='text-xs text-altwhite'>{t('settings.general.closeToTray.description')}</p>
+          </div>
+          <SettingsSwitch type='general' name='closeToTray' />
+        </div>
+
+        <Divider className='bg-border/70 my-4' />
+
+        <div className='flex justify-between items-center'>
+          <div className='flex flex-col gap-2 w-1/2'>
             <p className='text-sm text-content font-bold'>{t('settings.general.chatSounds')}</p>
             <p className='text-xs text-altwhite'>{sliderLabel}</p>
           </div>